+++ test/unit/mysql_rep_cmp/run_program.py
@@ -35,16 +35,41 @@ import version
 __version__ = version.__version__
 
 
-class Server2(object):
+class Mail(object):
 
-    """Class:  Server
+    """Class:  Mail
 
-    Description:  Class stub holder for mysql_class.Server class.
+    Description:  Class stub holder for gen_class.Mail class.
 
     Methods:
         __init__ -> Class initialization.
-        connect -> Method stub holder for mysql_class.Server.connect.
-        set_srv_binlog_crc -> Stub for mysql_class.Server.set_srv_binlog_crc.
+        add_2_msg -> Stub method holder for Mail.add_2_msg.
+        send_mail -> Stub method holder for Mail.send_mail.
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.data = None
+
+
+class SlaveRep(object):
+
+    """Class:  Server
+
+    Description:  Class stub holder for mysql_class.SlaveRep class.
+
+    Methods:
+        __init__ -> Class initialization.
+        connect -> Method stub holder for mysql_class.SlaveRep.connect.
 
     """
 
@@ -65,42 +90,36 @@ class Server2(object):
         self.do_tbl = {}
         self.ign_tbl = {}
         self.server_id = 11
+        self.conn_msg = None
 
-    def connect(self):
+    def connect(self, silent=False):
 
         """Method:  connect
 
-        Description:  Method stub holder for mysql_class.Server.connect.
+        Description:  Method stub holder for mysql_class.SlaveRep.connect.
 
         Arguments:
 
         """
 
-        return True
+        status = True
 
-    def show_slv_hosts(self):
+        if silent:
+            status = True
 
-        """Method:  show_slv_hosts
-
-        Description:  Stub holder for mysql_class.MasterRep.show_slv_hosts.
-
-        Arguments:
-
-        """
-
-        return [{"Server_id": 10}]
+        return status
 
 
-class Server(object):
+class MasterRep(object):
 
     """Class:  Server
 
-    Description:  Class stub holder for mysql_class.Server class.
+    Description:  Class stub holder for mysql_class.MasterRep class.
 
     Methods:
         __init__ -> Class initialization.
-        connect -> Method stub holder for mysql_class.Server.connect.
-        set_srv_binlog_crc -> Stub for mysql_class.Server.set_srv_binlog_crc.
+        connect -> Method stub holder for mysql_class.MasterRep.connect.
+        show_slv_hosts -> Metho for mysql_class.MasterRep.show_slv_hosts.
 
     """
 
@@ -121,18 +140,25 @@ class Server(object):
         self.do_tbl = {}
         self.ign_tbl = {}
         self.server_id = 10
+        self.slv_lists = [{"Server_id": 11}]
+        self.conn_msg = None
 
-    def connect(self):
+    def connect(self, silent=False):
 
         """Method:  connect
 
-        Description:  Method stub holder for mysql_class.Server.connect.
+        Description:  Method stub holder for mysql_class.MasterRep.connect.
 
         Arguments:
 
         """
 
-        return True
+        status = True
+
+        if silent:
+            status = True
+
+        return status
 
     def show_slv_hosts(self):
 
@@ -144,7 +170,7 @@ class Server(object):
 
         """
 
-        return [{"Server_id": 10}]
+        return self.slv_lists
 
 
 class UnitTest(unittest.TestCase):
@@ -155,6 +181,17 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
+        test_both_conn_fail -> Test with failed connection on master and slave.
+        test_slave_conn_fail -> Test with failed connection on slave.
+        test_master_conn_fail -> Test with failed connection on master.
+        test_conn_success -> Test with successful connection to master & slave.
+        test_str_server_id -> Test with integer server_id.
+        test_int_server_id -> Test with integer server_id.
+        test_no_std_out -> Test with no standard out suppression selected.
+        test_email_no_subj_mailx -> Test email using mailx and no subj line.
+        test_email_no_subj -> Test with email, but no subject line.
+        test_email_mailx -> Test with using mailx command.
+        test_email -> Test with email setup.
         test_slave_not_present -> Test with slave not in replic set.
         test_database_option -> Test with database option in args_array.
         test_table_option -> Test with table option in args_array.
@@ -172,14 +209,256 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.err_msg = "Failed Conection Message"
+        self.mail = Mail()
+        self.master = MasterRep()
+        self.slave = SlaveRep()
         self.sys_ign_db = ["performance_schema", "information_schema"]
         self.args_array = {"-c": True, "-d": True, "-r": True}
+        self.args_array2 = {"-c": True, "-d": True, "-r": True,
+                            "-e": "email_address", "-s": "subject_line"}
+        self.args_array2a = {"-c": True, "-d": True, "-r": True,
+                             "-e": "email_address", "-s": "subject_line",
+                             "-u": True}
+        self.args_array3 = {"-c": True, "-d": True, "-r": True,
+                            "-e": "email_address"}
+        self.args_array3a = {"-c": True, "-d": True, "-r": True,
+                             "-e": "email_address", "-u": True}
+        self.args_array4 = {"-c": True, "-d": True, "-r": True, "-z": True}
 
-    @mock.patch("mysql_rep_cmp.cmds_gen.disconnect",
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance",
-                mock.Mock(return_value=Server2()))
-    def test_slave_not_present(self):
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_both_conn_fail(self, mock_server):
+
+        """Function:  test_both_conn_fail
+
+        Description:  Test with failed connection on master and slave.
+
+        Arguments:
+
+        """
+
+        self.master.conn_msg = self.err_msg
+        self.slave.conn_msg = self.err_msg
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        with gen_libs.no_std_out():
+            self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
+                                                       self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_slave_conn_fail(self, mock_server):
+
+        """Function:  test_slave_conn_fail
+
+        Description:  Test with failed connection on slave.
+
+        Arguments:
+
+        """
+
+        self.slave.conn_msg = self.err_msg
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        with gen_libs.no_std_out():
+            self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
+                                                       self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_master_conn_fail(self, mock_server):
+
+        """Function:  test_master_conn_fail
+
+        Description:  Test with failed connection on master.
+
+        Arguments:
+
+        """
+
+        self.master.conn_msg = self.err_msg
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        with gen_libs.no_std_out():
+            self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
+                                                       self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_conn_success(self, mock_server):
+
+        """Function:  test_conn_success
+
+        Description:  Test with successful connection to master and slave.
+
+        Arguments:
+
+        """
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_str_server_id(self, mock_server):
+
+        """Function:  test_str_server_id
+
+        Description:  Test with string server_id.
+
+        Arguments:
+
+        """
+
+        self.master.slv_lists = [{"Server_id": "11"}]
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_int_server_id(self, mock_server):
+
+        """Function:  test_int_server_id
+
+        Description:  Test with integer server_id.
+
+        Arguments:
+
+        """
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_no_std_out(self, mock_server):
+
+        """Function:  test_no_std_out
+
+        Description:  Test with no standard out suppression selected.
+
+        Arguments:
+
+        """
+
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array4,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    @mock.patch("mysql_rep_cmp.gen_class.setup_mail")
+    def test_email_no_subj_mailx(self, mock_mail, mock_server):
+
+        """Function:  test_email_no_subj_mailx
+
+        Description:  Test with email using mailx and no subject line.
+
+        Arguments:
+
+        """
+
+        mock_mail.return_value = self.mail
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array3a,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    @mock.patch("mysql_rep_cmp.gen_class.setup_mail")
+    def test_email_no_subj(self, mock_mail, mock_server):
+
+        """Function:  test_email_no_subj
+
+        Description:  Test with email, but no subject line.
+
+        Arguments:
+
+        """
+
+        mock_mail.return_value = self.mail
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array3,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    @mock.patch("mysql_rep_cmp.gen_class.setup_mail")
+    def test_email_mailx(self, mock_mail, mock_server):
+
+        """Function:  test_email_mailx
+
+        Description:  Test with using mailx command.
+
+        Arguments:
+
+        """
+
+        mock_mail.return_value = self.mail
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array2a,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    @mock.patch("mysql_rep_cmp.gen_class.setup_mail")
+    def test_email(self, mock_mail, mock_server):
+
+        """Function:  test_email
+
+        Description:  Test with email setup.
+
+        Arguments:
+
+        """
+
+        mock_mail.return_value = self.mail
+        mock_server.side_effect = [self.master, self.slave]
+
+        self.assertFalse(mysql_rep_cmp.run_program(self.args_array2,
+                                                   self.sys_ign_db))
+
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
+                mock.Mock(return_value=True))
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_slave_not_present(self, mock_server):
 
         """Function:  test_slave_not_present
 
@@ -189,16 +468,19 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.slave.server_id = 12
+
+        mock_server.side_effect = [self.master, self.slave]
+
         with gen_libs.no_std_out():
             self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
                                                        self.sys_ign_db))
 
     @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.cmds_gen.disconnect",
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance",
-                mock.Mock(return_value=Server()))
-    def test_database_option(self):
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_database_option(self, mock_server):
 
         """Function:  test_database_option
 
@@ -208,17 +490,18 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_server.side_effect = [self.master, self.slave]
+
         self.args_array["-B"] = "db1"
 
         self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
                                                    self.sys_ign_db))
 
     @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.cmds_gen.disconnect",
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance",
-                mock.Mock(return_value=Server()))
-    def test_table_option(self):
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_table_option(self, mock_server):
 
         """Function:  test_table_option
 
@@ -228,6 +511,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_server.side_effect = [self.master, self.slave]
+
         self.args_array["-t"] = ["tbl1", "tbl2"]
         self.args_array["-B"] = "db1"
 
@@ -235,11 +520,10 @@ class UnitTest(unittest.TestCase):
                                                    self.sys_ign_db))
 
     @mock.patch("mysql_rep_cmp.setup_cmp", mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.cmds_gen.disconnect",
+    @mock.patch("mysql_rep_cmp.mysql_libs.disconnect",
                 mock.Mock(return_value=True))
-    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance",
-                mock.Mock(return_value=Server()))
-    def test_run_program(self):
+    @mock.patch("mysql_rep_cmp.mysql_libs.create_instance")
+    def test_run_program(self, mock_server):
 
         """Function:  test_run_program
 
@@ -248,6 +532,8 @@ class UnitTest(unittest.TestCase):
         Arguments:
 
         """
+
+        mock_server.side_effect = [self.master, self.slave]
 
         self.assertFalse(mysql_rep_cmp.run_program(self.args_array,
                                                    self.sys_ign_db))